requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Deprecated==1.2.14
Django==4.2.13
django-formtools==2.5.1
django-phone-verify==3.0.0
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
nexmo==2.5.2
phonenumbers==8.13.36
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
sqlparse==0.5.0
twilio==9.0.5
typing_extensions==4.11.0
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
